np-41723: rename version to today's date.
add karate test for new version
