+++ cristin-organization/src/main/java/no/unit/nva/cristin/organization/CristinOrgApiClientVersion2.java
@@ -17,8 +17,8 @@ import java.util.stream.Collectors;
 import no.unit.nva.cristin.common.client.ApiClient;
 import no.unit.nva.cristin.common.client.IQueryApiClient;
 import no.unit.nva.cristin.model.SearchResponse;
-import no.unit.nva.cristin.organization.dto.version20230510.mapper.OrganizationFromUnitMapper;
-import no.unit.nva.cristin.organization.dto.version20230510.UnitDto;
+import no.unit.nva.cristin.organization.dto.version20230526.mapper.OrganizationFromUnitMapper;
+import no.unit.nva.cristin.organization.dto.version20230526.UnitDto;
 import no.unit.nva.model.Organization;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
 import nva.commons.apigateway.exceptions.BadGatewayException;
